+++ compiler/src/main/scala/dx/translator/Extras.scala
@@ -43,11 +43,21 @@ object ExtrasJsonProtocol extends DefaultJsonProtocol {
             timeout("*").convertTo[DxTimeout]
           }
           val headJobOnDemand = JsUtils.getOptionalBoolean(fields, DxRunSpec.HeadJobOnDemand)
+          val release =
+            JsUtils.getOptionalString(fields, DxRunSpec.release).map {
+              case release if Constants.allowedOsReleases.contains(release) => release
+              case release =>
+                throw new Exception(
+                    s"Unsupported execution environment value ${release}." +
+                      s" Supported values are ${Constants.allowedOsReleases.mkString(", ")}"
+                )
+            }
           DxRunSpec(
               fields.get(DxRunSpec.Access).map(_.convertTo[DxAccess]),
               fields.get(DxRunSpec.ExecutionPolicy).map(_.convertTo[DxExecPolicy]),
               restartableEntryPoints,
               timeout,
+              release,
               headJobOnDemand
           )
         case _ =>
@@ -61,7 +71,8 @@ object ExtrasJsonProtocol extends DefaultJsonProtocol {
           runSpec.timeoutPolicy.map(x => DxRunSpec.TimeoutPolicy -> JsObject("*" -> x.toJson)),
           runSpec.executionPolicy.map(x => DxRunSpec.ExecutionPolicy -> x.toJson),
           runSpec.restartableEntryPoints.map(x => DxRunSpec.RestartableEntryPoints -> JsString(x)),
-          runSpec.headJobOnDemand.map(x => DxRunSpec.HeadJobOnDemand -> JsBoolean(x))
+          runSpec.headJobOnDemand.map(x => DxRunSpec.HeadJobOnDemand -> JsBoolean(x)),
+          runSpec.release.map(x => DxRunSpec.release -> JsString(x))
       ).flatten.toMap
       if (fields.isEmpty) {
         JsNull
@@ -263,6 +274,7 @@ case class DxRunSpec(access: Option[DxAccess],
                      executionPolicy: Option[DxExecPolicy],
                      restartableEntryPoints: Option[String],
                      timeoutPolicy: Option[DxTimeout],
+                     release: Option[String],
                      headJobOnDemand: Option[Boolean] = None) {}
 
 object DxRunSpec {
@@ -272,6 +284,7 @@ object DxRunSpec {
   val TimeoutPolicy = "timeoutPolicy"
   val EntryPointNames = Set("all", "master")
   val HeadJobOnDemand = "headJobOnDemand"
+  val release = "release"
 
   def toApiJson(runSpec: DxRunSpec): Map[String, JsValue] = {
     runSpec.toJson.asJsObject.fields.filterNot {