+++ executorWdl/src/test/scala/dx/executor/wdl/WdlWorkflowExecutorTest.scala
@@ -116,7 +116,7 @@ class WdlWorkflowExecutorTest extends AnyFlatSpec with Matchers {
         gpu = false,
         Vector(
             ExecutionEnvironment(Constants.OsDistribution,
-                                 Constants.OsRelease,
+                                 Constants.DefaultOsRelease,
                                  Vector(Constants.OsVersion))
         ),
         Some(DiskType.SSD),